+++ ttsrv/logger_test.go
@@ -1,4 +1,4 @@
-package tt
+package ttsrv
 
 import (
 	"bytes"
@@ -18,32 +18,36 @@ func ExampleLogger_In() {
 	l.SetRemote("1.2.3.4:0001")
 
 	l.In(ttx.NoopHandler)(nil, mq.Pub(0, "a/b", "gopher"))
+	l.In(ttx.NoopHandler)(nil, mq.NewConnect())
 
 	// output:
-	// in  PUBLISH ---- p0 a/b 14 bytes
+	// in  PUBLISH ---- p0 a/b 14 bytes <- 1.2.3.4:0001:
+	// in  CONNECT ---- -------- MQTT5  0s 15 bytes <- 1.2.3.4:0001:
 }
 
 func ExampleLogger_Out() {
 	l := NewLogger()
 	l.SetOutput(os.Stdout)
+	l.SetRemote("1.2.3.4:0001")
 
 	p := mq.Pub(0, "a/b", "gopher")
 	l.Out(ttx.NoopHandler)(nil, p)
 
 	// output:
-	// out PUBLISH ---- p0 a/b 14 bytes
+	// out PUBLISH ---- p0 a/b 14 bytes -> 1.2.3.4:0001:
 }
 
 func ExampleLogger_DumpPacket() {
 	l := NewLogger()
 	l.SetOutput(os.Stdout)
 	l.SetDebug(true)
+	l.SetRemote("1.2.3.4:0001")
 
 	p := mq.Pub(0, "a/b", "gopher")
 	l.In(ttx.NoopHandler)(nil, p)
 
 	// output:
-	// in  PUBLISH ---- p0 a/b 14 bytes
+	// in  PUBLISH ---- p0 a/b 14 bytes <- 1.2.3.4:0001:
 	// 00000000  30 0c 00 03 61 2f 62 00  67 6f 70 68 65 72        |0...a/b.gopher|
 }
 
@@ -53,23 +57,18 @@ func ExampleLogger_SetMaxIDLen() {
 	l.SetRemote("1.2.3.4:0001")
 	l.SetMaxIDLen(6)
 	{
-		p := mq.NewConnect()
-		p.SetClientID("short")
-		l.Out(ttx.NoopHandler)(nil, p)
-	}
-	{
 		p := mq.NewConnAck()
 		p.SetAssignedClientID("1bbde752-5161-11ed-a94b-675e009b6f46")
-		l.In(ttx.NoopHandler)(nil, p)
+		l.Out(ttx.NoopHandler)(nil, p)
 	}
 	// output:
-	// short out CONNECT ---- -------- MQTT5 short 0s 20 bytes
-	// ~9b6f46 in  CONNACK ---- -------- 1bbde752-5161-11ed-a94b-675e009b6f46 44 bytes
+	// out CONNACK ---- -------- 1bbde752-5161-11ed-a94b-675e009b6f46 44 bytes -> 1.2.3.4:0001:~9b6f46
 }
 
 func ExampleLogger_errors() {
 	l := NewLogger()
 	l.SetOutput(os.Stdout)
+	l.SetRemote("1.2.3.4:0001")
 
 	broken := func(context.Context, mq.Packet) error {
 		return fmt.Errorf("broken")
@@ -78,9 +77,9 @@ func ExampleLogger_errors() {
 	l.In(broken)(nil, p)
 	l.Out(broken)(nil, p)
 	// output:
-	// in  PUBLISH ---- p0 a/b 14 bytes
+	// in  PUBLISH ---- p0 a/b 14 bytes <- 1.2.3.4:0001:
 	// broken
-	// out PUBLISH ---- p0 a/b 14 bytes
+	// out PUBLISH ---- p0 a/b 14 bytes -> 1.2.3.4:0001:
 	// broken
 }
 
@@ -110,18 +109,18 @@ func TestLogger(t *testing.T) {
 	var buf bytes.Buffer
 	l.SetOutput(&buf)
 	cid := "1bbde752-5161-11ed-a94b-675e009b6f46"
-	p := mq.NewConnect()
-	p.SetClientID(cid)
+	p := mq.NewConnAck()
+	p.SetAssignedClientID(cid)
 
 	// trimmed client id
 	l.Out(ttx.NoopHandler)(nil, p)
-	if v := buf.String(); !strings.HasPrefix(v, "~75e009b6f46") {
+	if v := buf.String(); !strings.HasSuffix(v, "~75e009b6f46\n") {
 		t.Error(v)
 	}
 
 	// subsequent
 	l.Out(ttx.NoopHandler)(nil, p)
-	if v := buf.String(); !strings.HasPrefix(v, "~75e009b6f46") {
+	if v := buf.String(); !strings.HasSuffix(v, "~75e009b6f46\n") {
 		t.Error(v)
 	}
 
@@ -132,7 +131,7 @@ func TestLogger(t *testing.T) {
 
 	buf.Reset()
 	l.Out(ttx.NoopHandler)(nil, p)
-	if v := buf.String(); !strings.Contains(v, "|f46|") {
+	if v := buf.String(); !strings.Contains(v, "f46|") {
 		t.Error(v)
 	}
 }